requirements.txt
bluepy
paho-mqtt
adafruit-circuitpython-charlcd
RPLCD.i2c
